Improve types organisation
