+++ assets/js/blocks/migration-products-to-product-collection/index.ts
@@ -2,3 +2,4 @@ export * from './migration-from-products-to-product-collection';
 export * from './migration-from-product-collection-to-products';
 export * from './migration-utils';
 export * from './constants';
+export * from './types';
