+++ assets/js/blocks/migration-products-to-product-collection/types.ts
@@ -1,0 +1,26 @@
+/**
+ * External dependencies
+ */
+import { type BlockInstance } from '@wordpress/blocks';
+
+export type GetBlocksClientIds = ( blocks: BlockInstance[] ) => string[];
+export type IsBlockType = ( block: BlockInstance ) => boolean;
+export type TransformBlock = (
+	block: BlockInstance,
+	innerBlock: BlockInstance[]
+) => BlockInstance;
+export type ProductGridLayoutTypes = 'flex' | 'list';
+export type PostTemplateLayoutTypes = 'grid' | 'default';
+
+export type ProductGridLayout = {
+	type: ProductGridLayoutTypes;
+	columns: number;
+};
+
+export type PostTemplateLayout = {
+	type: PostTemplateLayoutTypes;
+	columnCount: number;
+};
+export type UpgradeNoticeStatus = {
+	status: 'notseen' | 'seen' | 'reverted';
+};
